updated code to run on flutter 1.22.0 and other linting changes

+++ nutrigram_app/lib/ui/views/nutrient_info_display/nutrient_info_display_view.dart
@@ -200,7 +200,7 @@ class NutrientInfoDisplayView extends StatelessWidget {
             return Column(
               children: [
                 DProgressBar(
-                  percent: ratio * 100,
+                  percent: (ratio * 100).toDouble(),
                   title: e.type,
                   color: e.color,
                   value: "${e.value.toString()} ${e.unit}",
